move search to handler
