+++ source/dotnet/wholesale-api/Batches/Batches.Application/BatchApplicationService.cs
@@ -33,56 +33,9 @@ public class BatchApplicationService : IBatchApplicationService
         _batchDtoMapper = batchDtoMapper;
     }
 
-    public async Task<IEnumerable<BatchDto>> SearchAsync(
-        IEnumerable<string> filterByGridAreaCodes,
-        BatchState? filterByExecutionState,
-        DateTimeOffset? minExecutionTime,
-        DateTimeOffset? maxExecutionTime,
-        DateTimeOffset? periodStart,
-        DateTimeOffset? periodEnd)
-    {
-        var executionStateFilter = filterByExecutionState switch
-        {
-            null => Array.Empty<BatchExecutionState>(),
-            BatchState.Pending => new[] { BatchExecutionState.Created, BatchExecutionState.Submitted, BatchExecutionState.Pending },
-            BatchState.Executing => new[] { BatchExecutionState.Executing },
-            BatchState.Completed => new[] { BatchExecutionState.Completed },
-            BatchState.Failed => new[] { BatchExecutionState.Failed },
-            _ => throw new ArgumentOutOfRangeException(nameof(filterByExecutionState)),
-        };
-
-        var gridAreaFilter = filterByGridAreaCodes
-            .Select(g => new GridAreaCode(g))
-            .ToList();
-
-        var minExecutionTimeStart = ConvertToInstant(minExecutionTime);
-        var maxExecutionTimeStart = ConvertToInstant(maxExecutionTime);
-        var periodStartInstant = ConvertToInstant(periodStart);
-        var periodEndInstant = ConvertToInstant(periodEnd);
-
-        var batches = await _batchRepository
-            .SearchAsync(
-                gridAreaFilter,
-                executionStateFilter,
-                minExecutionTimeStart,
-                maxExecutionTimeStart,
-                periodStartInstant,
-                periodEndInstant)
-            .ConfigureAwait(false);
-
-        return batches.Select(_batchDtoMapper.Map);
-    }
-
     public async Task<BatchDto> GetAsync(Guid batchId)
     {
         var batch = await _batchRepository.GetAsync(batchId).ConfigureAwait(false);
         return _batchDtoMapper.Map(batch);
-    }
-
-    private static Instant? ConvertToInstant(DateTimeOffset? dateTimeOffset)
-    {
-        return dateTimeOffset == null
-            ? null
-            : Instant.FromDateTimeOffset(dateTimeOffset.Value);
     }
 }
